refactor(release): move executor renderers
