+++ packages/release/src/api/executor/__.ts
@@ -8,10 +8,7 @@ export {
   type ExecutionGraphNode,
   type ExecutionResult,
   type ExecutionStatus,
-  formatExecutionStatus,
-  formatLifecycleEvent,
   type ObservableExecutionError,
-  type LifecycleEventLine,
   type ObservableExecutionRequirements,
   type ObservableResult,
   type ObservableResumeResult,
