+++ packages/release/src/api/executor/execute.test.ts
@@ -1,4 +1,3 @@
-import { Str } from '@kitz/core'
 import { Fs } from '@kitz/fs'
 import { Pkg } from '@kitz/pkg'
 import { Semver } from '@kitz/semver'
@@ -8,7 +7,7 @@ import { ReleaseCommit } from '../analyzer/models/commit.js'
 import { Official } from '../planner/models/item-official.js'
 import { Plan } from '../planner/models/plan.js'
 import { OfficialIncrement } from '../version/models/official-increment.js'
-import { formatExecutionStatus, formatLifecycleEvent, toPayload } from './execute.js'
+import { toPayload } from './execute.js'
 
 const makePackage = (scope: string) => ({
   scope,
@@ -33,51 +32,6 @@ const makeRelease = (
   })
 
 describe('executor execute helpers', () => {
-  test('formats lifecycle events into printable log lines', () => {
-    expect(formatLifecycleEvent({ _tag: 'ActivityStarted', activity: 'publish' } as any)).toEqual({
-      level: 'info',
-      message: '  › Starting: publish',
-    })
-    expect(formatLifecycleEvent({ _tag: 'ActivityCompleted', activity: 'publish' } as any)).toEqual(
-      {
-        level: 'info',
-        message: '✓ Completed: publish',
-      },
-    )
-    expect(
-      formatLifecycleEvent({
-        _tag: 'ActivityFailed',
-        activity: 'publish',
-        error: 'boom',
-      } as any),
-    ).toEqual({
-      level: 'error',
-      message: '✗ Failed: publish - boom',
-    })
-    expect(formatLifecycleEvent({ _tag: 'WorkflowStarted' } as any)).toBeUndefined()
-  })
-
-  test('renders colored lifecycle events and workflow status summaries', () => {
-    const completed = formatLifecycleEvent(
-      { _tag: 'ActivityCompleted', activity: 'publish' } as any,
-      { color: true },
-    )
-    const status = formatExecutionStatus(
-      {
-        state: 'not-started',
-        executionId: 'release-official:test',
-        lifecycle: 'official',
-        plannedPackages: ['@kitz/core'],
-      },
-      { color: true },
-    )
-
-    expect(completed?.message).toContain('\u001b[')
-    expect(Str.Visual.strip(completed?.message ?? '')).toContain('Completed: publish')
-    expect(status).toContain('\u001b[')
-    expect(Str.Visual.strip(status)).toContain('Run `release apply` to start the workflow.')
-  })
-
   test('builds workflow payloads in dependency order and normalizes commit data', async () => {
     const plan = Plan.make({
       lifecycle: 'official',
