+++ packages/release/src/api/renderer/__.ts
@@ -1,3 +1,4 @@
+export * from './execution.js'
 export * from './explain.js'
 export * from './forecast-markdown.js'
 export * from './graph.js'
